Unified. This is the way.

+++ raspberry_pi/setup.sh
@@ -1,4 +1,12 @@
 #!/bin/bash -e
+#
+# This bootstraps the unified application on a Raspberry Pi.
+#
+if [ -e seestar_alp ] || [ -e ~/seestar_alp ]; then
+    echo "ERROR: Existing seestar_alp directory detected."
+    echo "       You should run the raspberry_pi/update.sh script instead."
+    exit 255
+fi
 
 sudo apt-get update
 sudo apt-get install -y git python3-pip
@@ -7,24 +15,22 @@ git clone https://github.com/smart-underworld/seestar_alp.git
 cd  seestar_alp
 
 src_home=$(pwd)
-mkdir logs
+mkdir -p logs
 
-sed -i -e 's/127.0.0.1/0.0.0.0/g' device/config.toml
+if [ ! -e device/config.toml ]; then
+    sed -e 's/127.0.0.1/0.0.0.0/g' device/config.toml.example > device/config.toml
+fi
 
 sudo  pip install -r requirements.txt --break-system-packages
 
 cd raspberry_pi
-cat systemd/seestar_device.service | sed -e "s|/home/.*/seestar_alp|$src_home|g" > /tmp/seestar_device.service
-cat systemd/seestar_front.service | sed -e "s|/home/.*/seestar_alp|$src_home|g" > /tmp/seestar_front.service
-sudo mv /tmp/seestar*.service /etc/systemd/system
+cat systemd/seestar.service | sed -e "s|/home/.*/seestar_alp|$src_home|g" > /tmp/seestar.service
+sudo mv /tmp/seestar.service /etc/systemd/system
 
 sudo systemctl daemon-reload
 
-sudo systemctl enable seestar_device
-sudo systemctl enable seestar_front
-
-sudo systemctl start seestar_device
-sudo systemctl start seestar_front
+sudo systemctl enable seestar
+sudo systemctl start seestar
 
 cat <<_EOF
 |-------------------------------------|
@@ -37,7 +43,9 @@ cat <<_EOF
 |  ./seestar_alp/logs                 |
 |                                     |
 | Systemd logs can be viewed via      |
-| journalctl -u seestar_device        |
-| journalctl -u seestar_front         |
+| journalctl -u seestar               |
+|                                     |
+| Current status can be viewed via    |
+| systemctl status seestar            |
 |-------------------------------------|
 _EOF
